behaviour field is now in Entity
Previously all subclasses of Entity had their own behaviour field and
getBehaviour() and setBehaviour() methods. Now it is only the Entity
class that has this field.

<logic
